+++ src/(tabs)/Search/index.tsx
@@ -1,13 +1,14 @@
 import {SafeAreaView, StyleSheet, Text, View} from 'react-native';
-import {Colors} from '../../assets/color/Colors';
-import {useTheme} from '../util/ThemeContext';
-const Post = () => {
+import {Colors} from '../../../assets/color/Colors';
+import {useTheme} from '../../util/ThemeContext';
+
+const Search = () => {
   const {theme} = useTheme();
   const color = Colors[theme];
   return (
     <SafeAreaView
       style={[styles.container, {backgroundColor: color.background}]}>
-      <Text>Post Screen</Text>
+      <Text style={{color: color.text}}>Search Screen</Text>
     </SafeAreaView>
   );
 };
@@ -17,4 +18,5 @@ const styles = StyleSheet.create({
     flex: 1,
   },
 });
-export default Post;
+
+export default Search;
